Delete User Tip

+++ backend/controllers/tip.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Tip = require("../models/tipModel");
+const User = require("../models/userModel");
 
 // * CREATE TIP
 // * ROUTE: POST /api/tips
@@ -13,6 +14,7 @@ const createTip = asyncHandler(async (req, res) => {
 
   const tip = await Tip.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(tip);
@@ -27,6 +29,20 @@ const updateTip = asyncHandler(async (req, res, next) => {
   if (!tip) {
     res.status(400);
     throw new Error("Tip not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // CHECK FOR USER
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // LOGIN USER MATCHES TIP USER
+  if (tip.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedTip = await Tip.findByIdAndUpdate(req.params.id, req.body, {
@@ -47,6 +63,20 @@ const deleteTip = asyncHandler(async (req, res, next) => {
     throw new Error("Tip not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  // CHECK FOR USER
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // LOGIN USER MATCHES TIP USER
+  if (tip.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   await tip.deleteOne();
 
   res.status(200).json({ id: req.params.id });
@@ -63,8 +93,8 @@ const getTip = asyncHandler(async (req, res, next) => {
 // * ROUTE: GET /api/tips
 // * ACCESS: Private
 const getAllTips = asyncHandler(async (req, res, next) => {
-  const tips = await Tip.find();
-  res.status(200).json(tips);
+  const goals = await Tip.find({ user: req.user.id });
+  res.status(200).json(goals);
 });
 
 module.exports = {
